added some debugging print statements
Removed some code that was commented out and no longer used.  Added some print statements to help find a bug that this app has.  If an item in DiceRollListView is deleted, or if the entire diceRoll array is emptied, this change is saved ...however, ContentView still shows the diceRolls array with no changes...