Massive overhaul on how data is saved
Thanks to @NigelGee from the HackingWithSwift forums!  He revamped the FileManager-DocumentsDirectory to make it handle any kind of data.  I think this was from one of Paul Hudson's lessons.  I need to study it.

Also he created a DiceController.swift that acts as datacontroller and injected into the environment.  I didnt know you could do that as I have only seen the Environment work with CoreData.  I need a lot more study but these changes will help me learn even more!

Lastly, he modified the DiceRollListView to show the color of the dice that was rolled...he hard coded a color of Blue, but I extended DiceData to hold the lastRolledColor in a String then that is updated on each roll.